enable SendFeedback button in Classifier
- move SendFeedbackDialog to components/common
- extract 'createGitHubIssue' function
- code cleanup
