+++ src/components/ApplicationList/ApplicationList.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { SettingsListItem } from "../SettingsListItem";
 import { HelpListItem } from "../HelpListItem";
-import { SendFeedbackListItem } from "../SendFeedbackListItem";
 import { List } from "@mui/material";
+import { SendFeedbackListItem } from "components/common/SendFeedbackListItem";
 
 export const ApplicationList = () => {
   return (
