+++ src/components/common/SendFeedbackDialog/index.ts
@@ -1,0 +1,1 @@
+export { SendFeedbackDialog } from "./SendFeedbackDialog";
