+++ src/components/common/SendFeedbackListItem/SendFeedbackListItem.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import { ListItem, ListItemIcon, ListItemText } from "@mui/material";
+import FeedbackIcon from "@mui/icons-material/Feedback";
+import { useDialog } from "hooks";
+import { SendFeedbackDialog } from "components/common/SendFeedbackDialog";
+
+export const SendFeedbackListItem = () => {
+  const { onClose, onOpen, open } = useDialog();
+
+  return (
+    <>
+      <ListItem button onClick={onOpen}>
+        <ListItemIcon>
+          <FeedbackIcon />
+        </ListItemIcon>
+
+        <ListItemText primary="Send feedback" />
+      </ListItem>
+
+      <SendFeedbackDialog onClose={onClose} open={open} />
+    </>
+  );
+};
