+++ src/utils/createGitHubIssue.ts
@@ -1,0 +1,17 @@
+import { AlertType } from "types/AlertStateType";
+
+export const createGitHubIssue = (
+  title: string,
+  body: string,
+  alertType: AlertType = AlertType.Error
+) => {
+  const label = alertType === AlertType.Error ? "bug" : "help%20wanted";
+  const url =
+    "https://github.com/piximi/prototype/issues/new?title=" +
+    encodeURIComponent(title) +
+    "&labels=" +
+    label +
+    "&body=" +
+    encodeURIComponent(body);
+  window.open(url);
+};
